requirements.txt
Django==2.0.6
django-filter==1.1.0
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
html2text==2018.1.9
Markdown==2.6.11
Pillow==5.1.0
pytz==2018.4
sorl-thumbnail==12.4.1
